Migrate javascript tests to test folder
