+++ test/javascript/DoctorCard.test.js
@@ -3,9 +3,9 @@ import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { Provider } from 'react-redux';
 
-import store from '../redux/store';
+import store from '../../app/javascript/redux/store';
 
-import CommunityCard from '../components/CommunityCard';
+import DoctorCard from '../../app/javascript/components/DoctorCard';
 
 describe('rendered App', () => {
   let renderedComponent;
@@ -13,15 +13,15 @@ describe('rendered App', () => {
   beforeEach(() => {
     renderedComponent = render(
       <Provider store={store}>
-        <CommunityCard />
+        <DoctorCard />
       </Provider>,
     );
   });
 
-  it('CommunityCard has a container with class App', () => {
+  it('DoctorCard has a container with class App', () => {
     const { getByRole } = renderedComponent;
-    const button = getByRole('button', { name: '' });
-    expect(button).toBeInTheDocument();
+    const img = getByRole('img', { name: '' });
+    expect(img).toBeInTheDocument();
   });
 
   // it('has a container with class App', () => {
