+++ test/javascript/LineChart.test.js
@@ -3,9 +3,9 @@ import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { Provider } from 'react-redux';
 
-import store from '../redux/store';
+import store from '../../app/javascript/redux/store';
 
-import Doctor from '../containers/Doctor';
+import LineChart from '../../app/javascript/charts/line.rechart';
 
 describe('rendered App', () => {
   let renderedComponent;
@@ -13,23 +13,23 @@ describe('rendered App', () => {
   beforeEach(() => {
     renderedComponent = render(
       <Provider store={store}>
-        <Doctor />
+        <LineChart />
       </Provider>,
     );
   });
 
-  it('Doctor has a container with class App', () => {
-    const { getByRole } = renderedComponent;
-    const tablist = getByRole('tablist', { name: '' });
-    expect(tablist).toBeInTheDocument();
+  it('LineChart has a container with class App', () => {
+    const { container } = renderedComponent;
+    const appElement = container.querySelector('.recharts-wrapper');
+    expect(appElement).toBeInTheDocument();
   });
 
-  // it('has a container with class App', () => {
-  //   const { getByRole } = renderedComponent;
-  //   const tab = getByRole('tab', { name: 'My Friends' });
-  //   expect(tab).toBeInTheDocument();
-  // });
-  //
+  it('LineChart has a container with class App', () => {
+    const { container } = renderedComponent;
+    const appElement = container.querySelector('svg');
+    expect(appElement.classList).toContain('recharts-surface');
+  });
+
   // it('has a container with class App', () => {
   //   const { getByRole } = renderedComponent;
   //   const tab = getByRole('tab', { name: 'My Communities' });
